requirements.txt
streamlit 
python-dotenv 
mistralai 
pycountry 
pydantic
